// AS241S3_T15-FundoSmart/src/app/core/interfaces/fruit.ts
export interface Fruit {
  id?: number;
  nom_frut: string;
  kil_pri: number;
  kil_ter: number;
  kil_qui: number;
  kil_mad: number;
  can_pri: number;
  can_ter: number;
  can_qui: number;
  can_mad: number;
  pre_pri: number;
  pre_ter: number;
  pre_qui: number;
  pre_mad: number;
  fecha: string; 
}